[yy] merge master

+++ EasyDAL.Exchange/UserFacade/Join/WhereX.cs
@@ -92,7 +92,7 @@ namespace Yunyong.DataExchange.UserFacade.Join
             result.PageSize = option.PageSize;
             var paras = DC.GetParameters();
             var sql = DC.SqlProvider.GetSQL<M>(UiMethodEnum.JoinQueryPagingListAsync, result.PageIndex, result.PageSize);
-            result.TotalCount = await SqlHelper.ExecuteScalarAsync<long>(DC.Conn, sql[0], paras);
+            result.TotalCount = await SqlHelper.ExecuteScalarAsync<int>(DC.Conn, sql[0], paras);
             result.Data = (await SqlHelper.QueryAsync<M>(DC.Conn, sql[1], paras)).ToList();
             return result;
         }
